requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.2.1
dj-static==0.0.6
Django==1.7.1
django-appconf==0.6
django-compressor==1.4
djangorestframework==3.0.0
drf-nested-routers==0.9.0
easy-thumbnails==2.5
gunicorn==19.1.1
idna==2.6
Pillow==5.0.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
six==1.8.0
sqlparse==0.2.4
static3==0.5.1
stripe==1.79.1
urllib3==1.22
uWSGI==2.0.15
whitenoise==2.0.6
